+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [UNRELEASED]
+
+- added support for setting up proxy
+
 ## [0.61.1]
 
 - correctly adding the built-in skills to apps